+++ packages/cms/lib/modules/resource-representation-widgets/lib/fields.js
@@ -9,12 +9,20 @@ const fields = [
     name: 'displayType',
     label: 'Representation',
     type: 'select',
-    choices: [             
-
+    choices: [
       {
         'label': 'Idea Page (only for idea resource)',
         'value': 'idea-page',
-        'showFields' : ['showShareButtons', 'displayRanking', 'shareChannelsSelection', 'defaultImage'].concat(labels).concat(timeLabels)
+        'showFields' : [
+          'showShareButtons',
+          'displayRanking',
+          'shareChannelsSelection',
+          'defaultImage',
+          'hideUserLastName',
+          'hideVotes',
+          'hideStatus',
+          'shareTitle'
+        ].concat(labels).concat(timeLabels)
       },
       {
         'label': 'Title',
@@ -118,6 +126,59 @@ const fields = [
             label: "Whatsapp"
         },
     ]
+  },
+  {
+      name: 'hideUserLastName',
+      type: 'boolean',
+      label: 'Hide last name of the user who sent in the idea?',
+      choices: [
+          {
+              value: true,
+              label: "Yes",
+          },
+          {
+              value: false,
+              label: "No"
+          },
+      ],
+      def: false
+  },
+  {
+      name: 'hideVotes',
+      type: 'boolean',
+      label: 'Hide votes in controls?',
+      choices: [
+          {
+              value: true,
+              label: "Yes",
+          },
+          {
+              value: false,
+              label: "No"
+          },
+      ],
+      def: false
+  },
+  {
+      name: 'hideStatus',
+      type: 'boolean',
+      label: 'Hide status in controls?',
+      choices: [
+          {
+              value: true,
+              label: "Yes",
+          },
+          {
+              value: false,
+              label: "No"
+          },
+      ],
+      def: false
+  },
+  {
+    name: 'shareTitle',
+    type: 'string',
+    label: 'Title above share buttons (default: Deel dit voorstel)'
   }
 ].concat(
     ideaStates.map((state) => {